refactor: inline some helper methods
