+++ packages/cli-tools/bin/streamr-stream-list.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import { Option } from 'commander'
-import { list } from '../src/list'
+import EasyTable from 'easy-table'
 import { createClient } from '../src/client'
 import { createCommand } from '../src/command'
 
@@ -14,7 +14,7 @@ createCommand()
         .default('stream_get'))
     .option('--public-access', 'include publicly available streams')
     .option('--no-granted-access', 'exclude streams that user has directly granted permissions to')
-    .action((options: any) => {
+    .action(async (options: any) => {
         const query: any = {
             operation: options.operation,
             noConfig: true,
@@ -23,6 +23,14 @@ createCommand()
             grantedAccess: options.grantedAccess
         }    
         const client = createClient(options)
-        list(query, client)
+        const streams = await client.listStreams(query)
+        if (streams.length > 0) {
+            // @ts-expect-error: TODO: lastUpdated not officially part of stream object?
+            console.info(EasyTable.print(streams.map(({id, name, lastUpdated}) => ({
+                lastUpdated,
+                id,
+                name
+            }))))
+        }
     })
-    .parse()
+    .parseAsync()
