+++ packages/cli-tools/bin/streamr-stream-show.ts
@@ -1,5 +1,4 @@
 #!/usr/bin/env node
-import { show } from '../src/show'
 import { getStreamId } from './common'
 import { createClient } from '../src/client'
 import { createCommand } from '../src/command'
@@ -8,9 +7,15 @@ createCommand()
     .arguments('<streamId>')
     .description('show detailed information about a stream')
     .option('--include-permissions', 'include list of permissions (requires SHARE permission)')
-    .action((streamIdOrPath: string, options: any) => {
+    .action(async (streamIdOrPath: string, options: any) => {
         const streamId = getStreamId(streamIdOrPath, options)!
         const client = createClient(options)
-        show(streamId, options.includePermissions, client)
+        const stream = await client.getStream(streamId)
+        const obj = stream.toObject()
+        if (options.includePermissions) {
+            // @ts-expect-error permissions not on {}
+            obj.permissions = await stream.getPermissions()
+        }
+        console.info(JSON.stringify(obj, null, 2))
     })
-    .parse()
+    .parseAsync()
