--- packages/cli-tools/src/subscribe.ts
@@ -1,8 +1,0 @@
-import { StreamrClient } from 'streamr-client'
-
-export const subscribe = (streamId: string, streamPartition: number, client: StreamrClient): void => {
-    client.subscribe({
-        streamId,
-        streamPartition,
-    }, (message) => console.info(JSON.stringify(message)))
-}
